Prevent accidentally-absolute paths for external binaries
Previously, if ctx.attr.binary.label.package was empty (this can happen in toplevel targets, or external dependencies), the _binary_name function would return a path with a leading /, which messes up path lookups.
With this patch, the leading slash is prevented if the target package is the emtpy string.