Remove icon support
The [AppImage Type2 Spec][appimage-spec] specifies that the contained [AppDir][appdir-spec] may contain a  [.desktop file][desktop-spec]. To my knowledge this is only used with [appimaged][appimaged] and entirely optional.

The only reason for very basic icon file support to exist in rules_appimage is that the first implementation used [appimagetool][appimagetool] to build appimages, which makes using an icon file mandatory.

I don't know of anyone who actually used this feature.
Dropping this feature is long overdue and should slightly speed up builds as well as produce smaller appimage artifacts because we don't need to copy, compress, and store the png file.

[appdir-spec]: https://rox.sourceforge.net/desktop/AppDirs.html
[appimage-spec]: https://github.com/AppImage/AppImageSpec/blob/master/draft.md
[desktop-spec]: https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html
[appimaged]: https://docs.appimage.org/user-guide/run-appimages.html#integrating-appimages-into-the-desktop
[appimagetool]: https://github.com/AppImage/appimagetool