Use mksquashfs pseudofile definitions for AppImage build (#228)
This changes the way that appimage squashfs files are built with
rules_appimage. Instead of taring up all the files and then providing
that to mksquashfs we build a pseudo file definitions list that tells
mksquashfs where to get all the files from. This avoids the expensive
tar operation when inputs change (building the pseudofile defs is pretty
cheap).
The mksquashfs action does take longer now (likely because it has to do
thousands of `sh -c "cat {file}"` invocations) but overall the build
times are lower for large appimages and the cachability remote execution
ability should be better.

There was also an issue with the most recent MkAppDir implementation
where the tar was populated in a ThreadPool which was not deterministic,
so file order inside the .tar and .sqfs may change, leading to changed
artifact hashes. This is resolved with this.