Generate a custom MANIFEST to fix hermetic finding of _repo_map (#295)
Turns out MANIFEST files contain absolute paths. That means if you move
an appimage to a different computer it will not be able to find its
_repo_mapping runfile anymore.
To fix this, with this PR rules_appimage will generate its own MANIFEST
instead of taking the pregenerated one. All it really needs to contain
is the _repo_mapping line anyways. Everything else is already at the
expected path inside the appimage squashfs structure anyways.